requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
pandas==2.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
xlsxwriter==3.1.2


















